Use the old text/xsl content-type
XSLT 2.0 specifies that the correct content type for referring to a
stylesheet is application/xslt+xml. However, I guess some browsers don't
support that type and expect an older de-facto standard, text/xsl, which
is not registered with IANA. I'm hoping changing this will make more
browsers apply the stylesheet when loading the sample Atom feed.